livepatch.py: pass uuid_dir from GetFiles::post() to PaTch::base_dir
Instead of taking UUID as an argument to the methods of PaTch to
reconstruct uuid_dir, pass it as an argument to its contructor and
save it as the class member 'base_dir'. Prefer the name 'base_dir'
over 'uuid_dir' because this class has no notion of UUIDs, it just
needs a base directory to eventually run kpatch.

Because we create a separate PaTch object for every request, it makes
sense to deduplicate code by computing it once and saving it as part
of the PaTch object. It also promotes defensive programming as there's
a central place where this path is computed, rather than in multiple
methods spread all over.